Update Jenkinsfile stages for SonarQube and Selenium
Updated the 'SonarQube analysis' and 'Selenium test' stages in the Jenkinsfile. The previously commented out code has been replaced with new steps. For SonarQube, the new steps include a script that sets up the SonarQube environment and echoes "Run SonarQube Scanner". The actual command to run SonarQube analysis is commented out. For Selenium, the new steps include a script to run end-to-end tests.