Updated Jenkinsfile and package files for SonarQube integration
Updated the `Jenkinsfile` to use `npm run sonarqube` for SonarQube analysis. Modified `package-lock.json` and `package.json` to include `sonarqube-scanner` and its dependencies. Added a `sonarqube` script in `package.json` for easy SonarQube analysis. Updated several dependencies and added new ones in `package-lock.json` and `package.json`. Notably, `sonarqube-scanner` was added as a devDependency. The `ws` package was removed, and the `xml` package was added. These changes represent a significant update to the project's dependencies for improved functionality and performance.