Change s3path construction in Jenkinsfile
The Jenkinsfile has been updated to modify the construction of the `s3path` variable. Instead of appending the branch name to the path, it now sets `s3path` to root ("/"). This adjustment means all files will be pushed to the root directory of the S3 bucket, rather than being stored in a branch-named directory.